Throw exception when data not found

+++ src/main/java/com/featureswitch/featureswitch/service/userfeature/UserFeatureServiceImpl.java
@@ -3,6 +3,7 @@ package com.featureswitch.featureswitch.service.userfeature;
 import com.featureswitch.featureswitch.entity.FeatureEntity;
 import com.featureswitch.featureswitch.entity.UserEntity;
 import com.featureswitch.featureswitch.entity.UserFeatureEntity;
+import com.featureswitch.featureswitch.exceptions.DataNotFoundException;
 import com.featureswitch.featureswitch.repository.UserFeatureRepository;
 import com.featureswitch.featureswitch.service.feature.FeatureService;
 import com.featureswitch.featureswitch.service.user.UserService;
@@ -19,29 +20,33 @@ public class UserFeatureServiceImpl implements UserFeatureService{
     private final UserFeatureRepository userFeatureRepository;
 
     @Override
-    public UserFeatureEntity updateByUserEmailAndFeatureName(String userEmail, String featureName, boolean enable) {
-        UserEntity user = userService.getUserByEmail(userEmail);
-        if (user == null) {
-            log.info("User does not exist");
-            return null;
+    public UserFeatureEntity updateByUserEmailAndFeatureName(String userEmail, String featureName, boolean enable) throws DataNotFoundException {
+
+        UserEntity user;
+        try {
+            user = userService.getUserByEmail(userEmail);
+        } catch (DataNotFoundException ex) {
+            log.info("updateByUserEmailAndFeatureName(): User not found");
+            throw ex;
         }
 
-        FeatureEntity feature = featureService.getFeatureByName(featureName);
-        if (feature == null) {
-            log.info("Feature does not exist");
-            return null;
+        FeatureEntity feature;
+        try {
+            feature = featureService.getFeatureByName(featureName);
+        } catch (DataNotFoundException ex) {
+            log.info("updateByUserEmailAndFeatureName(): Feature not found");
+            throw ex;
         }
 
         UserFeatureEntity userFeature = userFeatureRepository.findByUserIdAndFeatureId(user.getUserId(), feature.getFeatureId());
-        if (userFeature != null) {
-            if (!enable) {
-                userFeatureRepository.delete(userFeature);
-                log.info("User is disabled");
-                return userFeature;
-            }
-        }
 
-        if (enable) {
+        if (userFeature != null && !enable) {
+            // If there is a record in the table and we want to disable it, delete it from the table
+            userFeatureRepository.delete(userFeature);
+            log.info("User is disabled");
+            return userFeature;
+        } else if (userFeature == null && enable) {
+            // If there is not a record in the table and we want to enable a feature, add it to the table
             UserFeatureEntity newUserFeature = new UserFeatureEntity();
             newUserFeature.setUserId(user.getUserId());
             newUserFeature.setFeatureId(feature.getFeatureId());
@@ -53,17 +58,22 @@ public class UserFeatureServiceImpl implements UserFeatureService{
     }
 
     @Override
-    public boolean userIsEnabled(String userEmail, String featureName) {
-        UserEntity user = userService.getUserByEmail(userEmail);
-        if ( user == null) {
-            log.info("User does not exist");
-            return false;
+    public boolean userIsEnabled(String userEmail, String featureName) throws DataNotFoundException{
+
+        UserEntity user;
+        try {
+            user = userService.getUserByEmail(userEmail);
+        } catch (DataNotFoundException ex) {
+            log.info("userIsEnabled(): User not found");
+            throw ex;
         }
 
-        FeatureEntity feature = featureService.getFeatureByName(featureName);
-        if (feature == null) {
-            log.info("Feature does not exist");
-            return false;
+        FeatureEntity feature;
+        try {
+            feature = featureService.getFeatureByName(featureName);
+        } catch (DataNotFoundException ex) {
+            log.info("userIsEnabled(): Feature not found");
+            throw ex;
         }
 
         UserFeatureEntity userFeature = userFeatureRepository.findByUserIdAndFeatureId(user.getUserId(), feature.getFeatureId());
